% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.lapply.R
\name{set.chunkSize}
\alias{set.chunkSize}
\title{set.chunkSize}
\usage{
set.chunkSize(x = .Machine$integer.max)
}
\arguments{
\item{x}{an 'integer' defining a sensible max chunk size
\rDefault : .Machine$integer.max}
}
\description{
Set the served chunkSize
}
