% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.lapply.R
\name{set.sensible.threads}
\alias{set.sensible.threads}
\title{set.sensible.threads}
\usage{
set.sensible.threads(x = min(1, parallel::detectCores() - 1))
}
\arguments{
\item{x}{a 'numeric' defining a sensible number of threads to run. 
Default : \code{min(1,parallel::detectCores()-1)}}
}
\description{
set.sensible.threads
}
